feat: add shared multer upload middleware
