+++ middleware/upload.js
@@ -1,0 +1,21 @@
+const multer = require('multer');
+const { uploadsDir } = require('../utils/image');
+
+const fileFilter = (req, file, cb) => {
+  const allowed = ['image/jpeg', 'image/png', 'image/heic', 'image/heif'];
+  if (allowed.includes(file.mimetype)) {
+    cb(null, true);
+  } else {
+    cb(new Error('Only JPG, PNG, or HEIC images are allowed'));
+  }
+};
+
+const upload = multer({
+  dest: uploadsDir,
+  fileFilter,
+  limits: {
+    fileSize: 10 * 1024 * 1024 // 10MB
+  }
+});
+
+module.exports = upload;
